bowser: Move hardware specific .rc files to /vendor/etc/init/hw
The reason these files are moved to /vendor/etc/init/hw and
not /vendor/etc/init is because 'init' scans the /vendor/etc/init folder
for vendor service .rc fragments to import. That would either result
in duplicate imports or out-of-order imports of these hardware
specific .rc files and that *may* have unintended consequences.

Change-Id: I14531da05e2a3a256f734f8de96e1f1969c753f9